bugs from refactoring
