+++ index.html
@@ -17,7 +17,7 @@
   </section>
   <section class="tutorials">
     <button onclick='document.getElementById("tutorial-content").hidden = !document.getElementById("tutorial-content").hidden'>Toggle Tutorial</button>
-    <ul id="tutorial-content">
+    <ul hidden id="tutorial-content">
       <li>double click to create vertices, additional double clicks to toggle final</li>
       <li>right click on vertices (or ctrl) drag to create edges</li>
       <li>right click on vertices or edge texts to inspect</li>
